refactor: Button 컴포넌트 타입수정 및 코드 개선
